hello-world

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 /**
  * main - a program that prints the size of various computers with printf function
+ *
  * Return: Always 0 (Success)
  */
 int main(void)
@@ -10,11 +11,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float e;
-
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
+	printf("Size of a char: %lu byte(s)\n", sizeof(a));
+	printf("Size of a int: %lu byte(s)\n", sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(d));
+	printf("Size of a float: %lu byte(s)\n", sizeof(e));
 	return (0)
 }
